+++ sentry-async-profiler/src/main/java/io/sentry/asyncprofiler/profiling/JavaContinuousProfiler.java
@@ -6,6 +6,7 @@ import static java.util.concurrent.TimeUnit.SECONDS;
 import io.sentry.DataCategory;
 import io.sentry.IContinuousProfiler;
 import io.sentry.ILogger;
+import io.sentry.IProfilingCanceledCallback;
 import io.sentry.IScopes;
 import io.sentry.ISentryExecutorService;
 import io.sentry.ISentryLifecycleToken;
@@ -285,6 +286,12 @@ public final class JavaContinuousProfiler
       }
     }
   }
+
+  @Override
+  public void registerProfilingCanceledCallback(@NotNull IProfilingCanceledCallback callback) {}
+
+  @Override
+  public void unregisterProfilingCanceledCallback(@NotNull IProfilingCanceledCallback callback) {}
 
   private void stop(final boolean restartProfiler) {
     try (final @NotNull ISentryLifecycleToken ignored = lock.acquire()) {